# api/openapi.json
{
  "components": {
    "schemas": {
...
      "FavoriteTrack": {
        "properties": {
          "artist": {
            "title": "Artist",
            "type": "string"
          },
          "duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          },
          "playMode": {
            "anyOf": [
              {
                "enum": [
                  "jukebox",
                  "autocanonizer"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Playmode"
          },
          "sourceType": {
            "anyOf": [
              {
                "enum": [
                  "youtube",
                  "soundcloud",
                  "bandcamp",
                  "upload"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcetype"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "tuningParams": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tuningparams"
          },
          "uniqueSongId": {
            "title": "Uniquesongid",
            "type": "string"
          }
        },
        "required": [
          "uniqueSongId",
          "title",
          "artist"
        ],
        "title": "FavoriteTrack",
        "type": "object"
      },
      "FavoritesSyncPayload": {
        "properties": {
          "favorites": {
            "items": {
              "$ref": "#/components/schemas/FavoriteTrack"
            },
            "title": "Favorites",
            "type": "array"
          }
        },
        "required": [
          "favorites"
        ],
        "title": "FavoritesSyncPayload",
        "type": "object"
      },
      "FavoritesSyncRequest": {
        "properties": {
          "favorites": {
            "items": {
              "$ref": "#/components/schemas/FavoriteTrack"
            },
            "title": "Favorites",
            "type": "array"
          }
        },
        "required": [
          "favorites"
        ],
        "title": "FavoritesSyncRequest",
        "type": "object"
      },
      "FavoritesSyncResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "favorites": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FavoriteTrack"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Favorites"
          }
        },
        "required": [
          "code",
          "count"
        ],
        "title": "FavoritesSyncResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "The Forever Jukebox Analysis API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/audio/{job_id}": {
      "get": {
        "operationId": "get_audio_api_audio__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Audio"
      }
    },
    "/api/favorites/sync": {
      "post": {
        "operationId": "create_favorites_sync_api_favorites_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoritesSyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoritesSyncResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Create Favorites Sync"
      }
    },
    "/api/favorites/sync/{code}": {
      "get": {
        "operationId": "get_favorites_sync_api_favorites_sync__code__get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoritesSyncPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Favorites Sync"
      },
      "put": {
        "operationId": "update_favorites_sync_api_favorites_sync__code__put",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoritesSyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoritesSyncResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Favorites Sync"
      }
    },
...
    "/api/logs/{job_id}": {
      "get": {
        "operationId": "get_job_log_api_logs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Log"
      }
    },
...
  }
}